fix: clean up
