--- packages/data-fetch/src/eas/viem/eas.ts
@@ -1,1 +1,0 @@
-// TODO constract eas
